@@ -1,128 +1,97 @@
 {
   "name": "neuralweb",
   "version": "1.0.0",
-  "description": "Neuralweb app",
+  "description": "",
   "main": "index.js",
   "homepage": "https://neuralweb-app.ioak.io",
   "scripts": {
-    "start": "tsc && webpack serve --env target=development --config webpack.config.js --mode development",
-    "start:fast": "webpack serve --env target=development --config webpack.config.js --mode development",
-    "build:local": "tsc -w",
-    "build": "tsc && webpack --env target=production --config webpack.config.js --mode production",
-    "deploy": "npm run build&&gh-pages -d build",
-    "lint": "eslint . --ext .js,.jsx,.ts,.tsx",
-    "lint:fix": "eslint . --ext .js,.jsx,.ts,.tsx --fix",
-    "graphql:generate": "graphql-codegen --config codegen.yml"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "webpack serve --config webpack.config.js --mode development",
+    "build": "webpack --env target=production --config webpack.config.js --mode production",
+    "deploy": "npm run build&&gh-pages -d build"
   },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/ioak-dev/neuralweb.git"
+  },
+  "keywords": [],
   "author": "",
   "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/ioak-dev/neuralweb/issues"
+  },
   "devDependencies": {
-    "@babel/core": "^7.12.10",
-    "@babel/preset-env": "^7.12.11",
-    "@babel/preset-react": "^7.12.10",
-    "@babel/preset-typescript": "^7.12.7",
-    "@graphql-codegen/cli": "^1.21.4",
-    "@graphql-codegen/introspection": "^1.18.2",
-    "@graphql-codegen/typescript": "^1.22.0",
-    "@graphql-codegen/typescript-operations": "^1.17.16",
-    "@graphql-codegen/typescript-react-apollo": "^2.2.4",
-    "@typescript-eslint/eslint-plugin": "^4.14.1",
-    "@typescript-eslint/parser": "^4.14.1",
-    "autoprefixer": "^10.4.7",
-    "babel-loader": "^8.2.2",
-    "css-loader": "^5.0.1",
-    "dotenv": "^8.2.0",
-    "eslint": "^7.18.0",
-    "eslint-config-airbnb": "^18.2.1",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-prettier": "^3.3.1",
-    "eslint-plugin-react": "^7.22.0",
-    "eslint-plugin-react-hooks": "^4.2.0",
+    "@babel/core": "^7.20.7",
+    "@babel/preset-env": "^7.20.2",
+    "@babel/preset-react": "^7.18.6",
+    "@babel/preset-typescript": "^7.18.6",
+    "@fortawesome/fontawesome-svg-core": "^6.2.1",
+    "@fortawesome/free-solid-svg-icons": "^6.2.1",
+    "@fortawesome/react-fontawesome": "^0.2.0",
+    "@material-ui/core": "^4.12.4",
+    "@material-ui/icons": "^4.11.3",
+    "@oakui/core-stage": "file:oakui-core-stage-1.0.0.tgz",
+    "@popperjs/core": "^2.11.6",
+    "@redux-devtools/core": "^3.13.1",
+    "@reduxjs/toolkit": "^1.9.1",
+    "@types/lodash": "^4.14.191",
+    "@types/react": "^18.0.26",
+    "@types/react-dom": "^18.0.10",
+    "@types/uuid": "^9.0.0",
+    "babel-loader": "^9.1.0",
+    "chart.js": "^3.9.1",
+    "chartjs-adapter-date-fns": "^2.0.0",
+    "cropperjs": "^1.5.13",
+    "css-loader": "^6.7.3",
+    "dotenv": "^16.0.3",
     "file-loader": "^6.2.0",
-    "gh-pages": "^3.1.0",
-    "html-webpack-plugin": "^5.3.1",
-    "postcss": "^8.4.13",
-    "postcss-loader": "^6.2.1",
-    "style-loader": "^2.0.0",
-    "tailwindcss": "^3.0.24",
-    "webpack": "^5.18.0",
-    "webpack-cli": "^4.4.0",
-    "webpack-dev-server": "^3.11.2"
+    "html-webpack-plugin": "^5.5.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-facebook-login": "^4.1.1",
+    "react-google-login": "^5.2.2",
+    "react-redux": "^8.0.5",
+    "react-router-dom": "^6.6.1",
+    "rxjs": "^7.8.0",
+    "sass": "^1.57.1",
+    "sass-loader": "^13.2.0",
+    "style-loader": "^3.3.1",
+    "tinycolor2": "^1.5.2",
+    "ts-loader": "^9.4.2",
+    "typescript": "^4.9.4",
+    "webpack": "^5.75.0",
+    "webpack-cli": "^5.0.1",
+    "webpack-dev-server": "^4.11.1"
   },
   "dependencies": {
-    "@apollo/client": "^3.3.16",
-    "@apollo/react-common": "^3.1.4",
-    "@apollo/react-components": "^4.0.0",
-    "@apollo/react-hoc": "^4.0.0",
-    "@codemirror/lang-markdown": "^0.20.0",
-    "@codemirror/language-data": "^0.20.0",
-    "@fortawesome/fontawesome-svg-core": "^6.1.1",
-    "@fortawesome/free-brands-svg-icons": "^6.1.1",
-    "@fortawesome/free-regular-svg-icons": "^6.1.1",
-    "@fortawesome/free-solid-svg-icons": "^6.1.1",
-    "@fortawesome/react-fontawesome": "^0.1.18",
-    "@material-ui/core": "^4.11.3",
-    "@material-ui/icons": "^4.11.2",
-    "@oakui/core-stage": "file:oakui-core-stage-1.0.0.tgz",
-    "@popperjs/core": "^2.11.5",
-    "@types/chart.js": "^2.9.36",
-    "@types/codemirror": "^5.60.5",
-    "@types/d3": "^7.1.0",
-    "@types/jest": "^26.0.20",
-    "@types/lodash": "^4.14.181",
-    "@types/marked": "^4.0.3",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
-    "@types/react-redux": "^7.1.16",
-    "@types/react-router": "^5.1.11",
-    "@types/react-router-dom": "^5.1.7",
-    "@types/rellax": "^1.7.4",
-    "@types/uuid": "^8.3.0",
-    "apollo-boost": "^0.4.9",
-    "apollo-cache-inmemory": "^1.6.6",
-    "apollo-client": "^2.6.10",
-    "apollo-link": "^1.2.14",
-    "apollo-link-context": "^1.0.20",
-    "apollo-link-error": "^1.1.13",
-    "apollo-link-http": "^1.5.17",
-    "axios": "^0.21.1",
-    "chart.js": "^3.7.1",
-    "chartjs-adapter-date-fns": "^2.0.0",
-    "codemirror": "^5.65.3",
-    "cropperjs": "^1.5.11",
-    "cross-env": "^7.0.3",
-    "d3": "^7.4.4",
-    "date-fns": "^2.16.1",
-    "focus-trap-react": "^8.11.1",
-    "graphql": "^15.5.0",
-    "graphql-tag": "^2.12.3",
-    "lodash": "^4.17.21",
-    "marked": "^4.0.15",
-    "prettier": "^2.2.1",
-    "query-string": "^7.0.0",
-    "quill": "^1.3.7",
-    "react": "^17.0.2",
-    "react-app-polyfill": "^2.0.0",
-    "react-chartjs-2": "^4.1.0",
-    "react-cookie": "^4.0.3",
-    "react-dom": "^17.0.2",
-    "react-notifications-component": "^3.0.3",
-    "react-popper": "^2.2.5",
-    "react-quill": "^1.3.5",
-    "react-redux": "^7.2.2",
-    "react-router": "^5.2.0",
-    "react-router-dom": "^5.2.0",
-    "redux": "^4.0.5",
-    "redux-thunk": "^2.3.0",
-    "rellax": "^1.12.1",
-    "rxjs": "^6.6.3",
-    "sass": "^1.56.2",
-    "sass-lint": "^1.13.0",
-    "sass-loader": "^13.2.0",
-    "tinycolor2": "^1.4.2",
-    "typescript": "^4.1.3",
-    "uuid": "^8.3.2"
+    "@tiptap/extension-color": "^2.0.0-beta.220",
+    "@tiptap/extension-font-family": "^2.0.0-beta.220",
+    "@tiptap/extension-highlight": "^2.0.0-beta.220",
+    "@tiptap/extension-image": "^2.0.0-beta.220",
+    "@tiptap/extension-link": "^2.0.0-beta.220",
+    "@tiptap/extension-subscript": "^2.0.0-beta.220",
+    "@tiptap/extension-superscript": "^2.0.0-beta.220",
+    "@tiptap/extension-table": "^2.0.0-beta.220",
+    "@tiptap/extension-table-cell": "^2.0.0-beta.220",
+    "@tiptap/extension-table-header": "^2.0.0-beta.220",
+    "@tiptap/extension-table-row": "^2.0.0-beta.220",
+    "@tiptap/extension-task-item": "^2.0.0-beta.220",
+    "@tiptap/extension-task-list": "^2.0.0-beta.220",
+    "@tiptap/extension-text-align": "^2.0.0-beta.220",
+    "@tiptap/extension-text-style": "^2.0.0-beta.220",
+    "@tiptap/extension-typography": "^2.0.0-beta.220",
+    "@tiptap/extension-underline": "^2.0.0-beta.220",
+    "@tiptap/extension-youtube": "^2.0.0-beta.220",
+    "@tiptap/pm": "^2.0.0-beta.220",
+    "@tiptap/react": "^2.0.0-beta.220",
+    "@tiptap/starter-kit": "^2.0.0-beta.220",
+    "authlite": "file:authlite-1.0.0.tgz",
+    "axios": "^1.2.2",
+    "basicui": "file:basicui-1.0.0.tgz",
+    "d3": "^7.8.4",
+    "file-saver": "^2.0.5",
+    "moment": "^2.29.4",
+    "react-popper": "^2.3.0",
+    "writeup": "file:writeup-1.0.0.tgz"
   }
 }
