@@ -1,96 +1,44 @@
-import {
-  faCheck,
-  faCircleNodes,
-  faFileExport,
-  faFilter,
-  faGlasses,
-  faLink,
-  faPen,
-  faPlus,
-  faTimes,
-} from '@fortawesome/free-solid-svg-icons';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import React, { useEffect, useRef, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { useHistory } from 'react-router';
 import './style.scss';
 import Topbar from '../../../components/Topbar';
-import Editor from '../../../components/Editor';
 import NoteModel from '../../../model/NoteModel';
-import Viewer from '../../../components/Viewer';
-import FileExplorer from '../../../components/FileExplorer';
-import ContextContent from '../../../components/ContextContent';
-import DisableContextBarCommand from '../../../events/DisableContextBarCommand';
-import Footer from '../../../components/Footer';
-import { getNoteByReference, saveNote } from './service';
-import NotelinkModel from '../../../model/NotelinkModel';
-import LinkView from '../../../components/LinkView';
+import { generateReport, getNoteByReference, saveNote } from './service';
 import GraphView from '../../../components/GraphView';
+import { useParams } from 'react-router-dom';
+import MainSection from '../../../components/MainSection';
+import ContentSection from '../../../components/Note/ContentSection';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { faCircleNodes, faDownload, faFile, faFileAlt, faHome, faLink } from '@fortawesome/free-solid-svg-icons';
+import { replaceNotelinkAutoItems } from '../../../store/actions/NotelinkAutoActions';
+import { updateNoteItem } from '../../../store/actions/NoteActions';
+import { isEmptyAttributes, isEmptyOrSpaces } from '../../../components/Utils';
+import { formatDate, formatDateText } from '../../../components/Lib/DateUtils';
 
 interface Props {
   location: any;
   space: string;
 }
 
-const queryString = require('query-string');
-
 const NotePage = (props: Props) => {
-  const history = useHistory();
-  const profile = useSelector((state: any) => state.profile);
-  const authorization = useSelector((state: any) => state.authorization);
   const dispatch = useDispatch();
-  const id = useRef('');
-
-  const [view, setView] = useState('view');
-  const [isContextExpanded, setIsContextExpanded] = useState(false);
-
-  const [state, setState] = useState<NoteModel | null>(null);
-
-  const [queryParam, setQueryParam] = useState<any>({});
+  const authorization = useSelector((state: any) => state.authorization);
+  const params = useParams();
+  const [note, setNote] = useState<NoteModel | null>(null);
+  const [view, setView] = useState<'home' | 'graph'>('home');
 
   useEffect(() => {
-    DisableContextBarCommand.next(false);
-  }, []);
-
-  useEffect(() => {
-    const queryParam = queryString.parse(props.location.search);
-    setQueryParam(queryParam);
-    id.current = queryParam?.id || '';
-  }, [props.location]);
-
-  useEffect(() => {
-    if (queryParam?.id && authorization.isAuth) {
-      setView('view');
-      getNoteByReference(props.space, queryParam?.id, authorization).then(
+    if (params.id && authorization.isAuth) {
+      getNoteByReference(props.space, params.id, authorization).then(
         (response: any) => {
-          setState(response);
-          setView('view');
+          if (!isEmptyAttributes(response)) {
+            setNote(response);
+          }
         }
       );
     }
-  }, [queryParam, authorization]);
+  }, [params, authorization]);
 
-  const handleChange = (_state: NoteModel) => {
-    setState(_state);
-  };
-
-  const openGraphMode = () => {
-    setView('graph');
-  };
-
-  const openEditMode = () => {
-    setView('edit');
-  };
-
-  const openViewMode = () => {
-    setView('view');
-  };
-
-  const closeNote = () => {
-    setView('view');
-    setState(null);
-    history.push(`/${props.space}/note`);
-  };
 
   // const saveChanges = () => {
   //   saveNote(props.space, state, authorization).then((response: any) => {
@@ -101,112 +49,52 @@ const NotePage = (props: Props) => {
   //   });
   // };
 
-  const handleSave = (_note: NoteModel) => {
-    saveNote(props.space, _note, authorization).then((response: any) => {
-      if (response && id.current === response.reference) {
-        setState(response);
-      }
-    });
-  };
-
-  // useEffect(() => {
-  //   console.log(state.content);
-  // }, [state.content]);
+  const onPostNoteSave = (data: any) => {
+    setNote(data.note);
+    dispatch(updateNoteItem({
+      ...data.note,
+      summary: isEmptyOrSpaces(data.note.summary) ? data.note.autoGeneratedSummary : data.note.summary
+    }));
+    dispatch(replaceNotelinkAutoItems({
+      newItems: data.notelinks,
+      noteRef: data.note.reference
+    }));
+  }
 
   return (
     <>
-      <div
-        className={`note-page page-animate ${
-          isContextExpanded
-            ? 'note-page--context-active'
-            : 'note-page--context-inactive'
-        }`}
-      >
-        <Topbar
-          title={state?.name || 'Untitled'}
-          fixed
-          isContextExpanded={isContextExpanded}
-        >
+      <div className='note-page page-animate'>
+        <Topbar title="Note" space={props.space}>
           <div className="topbar-actions">
             <button
-              className={`button ${view === 'view' ? 'active' : ''}`}
-              onClick={openViewMode}
+              className={`button ${view === 'home' ? 'active' : ''}`}
+              onClick={() => setView('home')}
             >
-              <FontAwesomeIcon icon={faGlasses} />
-              <span className="menu-highlight-line" />
-            </button>
-            <button
-              className={`button ${view === 'edit' ? 'active' : ''}`}
-              onClick={openEditMode}
-            >
-              <FontAwesomeIcon icon={faPen} />
+              <FontAwesomeIcon icon={faFileAlt} />
               <span className="menu-highlight-line" />
             </button>
             <button
               className={`button ${view === 'graph' ? 'active' : ''}`}
-              onClick={openGraphMode}
+              onClick={() => setView('graph')}
             >
               <FontAwesomeIcon icon={faCircleNodes} />
               <span className="menu-highlight-line" />
             </button>
-            <button
-              className={`button ${isContextExpanded ? 'active' : ''}`}
-              onClick={() => setIsContextExpanded(!isContextExpanded)}
-            >
-              <FontAwesomeIcon icon={faLink} />
-            </button>
           </div>
         </Topbar>
-        <div className="note-page__left">
-          <div className="note-page__main">
-            {view === 'edit' && state && (
-              <div className="main">
-                <Editor
-                  space={props.space}
-                  note={state}
-                  handleChange={handleChange}
-                  handleSave={handleSave}
-                />
-              </div>
-            )}
-            {view === 'view' && state && (
-              <div className="viewer-container main">
-                <Viewer space={props.space} note={state} />
-              </div>
-            )}
-            {view === 'graph' && state && (
-              <GraphView
-                space={props.space}
-                noteref={state.reference}
-                isContextExpanded={isContextExpanded}
-              />
-            )}
-            {['view', 'edit'].includes(view) && !queryParam.id && (
-              <div className="main">No file is open</div>
-            )}
-          </div>
-        </div>
+        {/* <MainSection> */}
+        {note && view === 'home' && <MainSection>
+          <ContentSection note={note} space={props.space} onPostNoteSave={onPostNoteSave} />
+        </MainSection>}
+        {note && view === 'graph' &&
+          <GraphView
+            space={props.space}
+            noteref={note.reference}
+            isContextExpanded={false}
+          />}
+        {/* </MainSection> */}
+
       </div>
-      <ContextContent space={props.space} isExpanded={isContextExpanded}>
-        {state && (
-          <div className="note-page__context">
-            <LinkView
-              space={props.space}
-              note={state}
-              handleClose={() => setIsContextExpanded(false)}
-            />
-          </div>
-        )}
-      </ContextContent>
-      {/* {view === 'edit' && state && (
-        <Footer>
-          <div className="footer-action">
-            <button className="button" onClick={saveChanges}>
-              <FontAwesomeIcon icon={faCheck} /> Save
-            </button>
-          </div>
-        </Footer>
-      )} */}
     </>
   );
 };
