@@ -1,8 +1,15 @@
 export default interface NoteModel {
-  _id: string;
+  _id?: string;
   name: string;
   reference: string;
-  content: string;
-  folderId: string;
-  color?: string;
+  content?: string;
+  contentText?: string;
+  summary?: string;
+  autoGeneratedSummary?: string;
+  labels: string[];
+  primaryLabel?: string;
+  keywords?: string[];
+  type?: 'Fleeting' | 'Reference' | 'Literature' | 'Permanent';
+  createdAt?: any;
+  [key: string]: any;
 }
